Refactoring
- Refactor PostLinkable a little bit (now value is not a dynamically typed
object but has one on few concrete types so it will be harder to make
a mistake when using it).
- Refactor ArchiveManager a little bit. Make ArchiveDescriptors have
databaseId and also all posts from loaded from archives have archive
descriptors so it's easy to figure out which post was loaded from an
archive and which is from 4chan.
- When loading posts for a thread first load posts from the cache and
only the check the database.
- Couple of fixes for the logic that determines which archived posts we
should retain when merging fresh server posts with archive posts. Now,
archive post is more valuable to us if it has more images than the
fresh server post.

+++ Kuroba/app/src/main/java/com/github/adamantcheese/chan/core/site/parser/PostParseCallable.java
@@ -74,7 +74,7 @@ class PostParseCallable implements Callable<Post> {
 
         return parser.parse(currentTheme, post, new PostParser.Callback() {
             @Override
-            public boolean isSaved(int postNo) {
+            public boolean isSaved(long postNo) {
                 return savedReplyManager.isSaved(post.board, postNo);
             }
