+++ Kuroba/app/src/main/java/com/github/adamantcheese/chan/ui/controller/settings/ThemeSettingsController.java
@@ -172,7 +172,7 @@ public class ThemeSettingsController
 
     private PostParser.Callback parserCallback = new PostParser.Callback() {
         @Override
-        public boolean isSaved(int postNo) {
+        public boolean isSaved(long postNo) {
             return false;
         }
